fixes issues with integration with rrdv6
