+++ lib/ts/recipe/thirdpartyemailpassword/thirdpartyEmailpasswordAuth.tsx
@@ -17,7 +17,7 @@
  * Imports.
  */
 import * as React from "react";
-import { PureComponent, useRef } from "react";
+import { PureComponent } from "react";
 
 import ThirdPartyEmailPassword from "./recipe";
 import { FeatureBaseProps } from "../../types";
@@ -73,17 +73,14 @@ export default function ThirdPartyAuthWrapper({
     onSessionExpired?: () => void;
 }): JSX.Element {
     const reactRouterDom = SuperTokens.getInstanceOrThrow().getReactRouterDom();
-    const Component = useRef<typeof ThirdPartyEmailPasswordAuth>(
-        reactRouterDom === undefined
-            ? ThirdPartyEmailPasswordAuth
-            : reactRouterDom.withRouter(ThirdPartyEmailPasswordAuth)
-    );
+    const history = reactRouterDom === undefined ? undefined : reactRouterDom.useHistoryCustom();
     return (
-        <Component.current
+        <ThirdPartyEmailPasswordAuth
+            history={history}
             onSessionExpired={onSessionExpired}
             requireAuth={requireAuth}
             recipe={ThirdPartyEmailPassword.getInstanceOrThrow()}>
             {children}
-        </Component.current>
+        </ThirdPartyEmailPasswordAuth>
     );
 }
